♻️  Refactoring(assistant): move and cleanup old chat code
