+++ fai-rag-app/fai-backend/fai_backend/llm/service.py
@@ -4,13 +4,9 @@ from langstream import Stream, join_final_output, as_async_generator
 from langstream.contrib import OpenAIChatStream, OpenAIChatMessage, OpenAIChatDelta
 
 from fai_backend.chat.stream import create_chat_stream_from_prompt
-from fai_backend.chat.template import CHAT_PROMPT_TEMPLATE_ARGS, SCORING_PROMPT_TEMPLATE_ARGS, chatPromptTemplate
-from fai_backend.config import settings
-from fai_backend.llm.impl.openai import OpenAILLM
-from fai_backend.llm.impl.parrot import ParrotLLM
-from fai_backend.llm.protocol import ILLMStreamProtocol
-from fai_backend.vector.service import VectorService
+from fai_backend.chat.template import CHAT_PROMPT_TEMPLATE_ARGS, SCORING_PROMPT_TEMPLATE_ARGS
 from fai_backend.vector.factory import vector_db
+from fai_backend.vector.service import VectorService
 
 SYSTEM_TEMPLATE = "You are a helpful AI assistant that helps people with answering questions about planning "
 "permission.<br> If you can't find the answer in the search result below, just say (in Swedish) "
@@ -18,15 +14,6 @@ SYSTEM_TEMPLATE = "You are a helpful AI assistant that helps people with answeri
 "question is not related to the context, politely respond that you are tuned to only "
 "answer questions that are related to the context.<br> The questions are going to be "
 "asked in Swedish. Your response must always be in Swedish."
-
-
-class LLMFactory:
-    @staticmethod
-    def get(backend: str = settings.LLM_BACKEND) -> ILLMStreamProtocol:
-        return {
-            "parrot": lambda: ParrotLLM(),
-            "openai": lambda: OpenAILLM(template=chatPromptTemplate),
-        }[backend]()
 
 
 async def query_vector(vector_service, collection_name, query, n_results=10):
