+++ fai-rag-app/fai-backend/fai_backend/main.py
@@ -6,17 +6,16 @@ from fastapi.middleware.cors import CORSMiddleware
 from sse_starlette import EventSourceResponse, ServerSentEvent
 from starlette.responses import HTMLResponse, RedirectResponse
 
+from fai_backend.assistant.routes import router as templates_router
 from fai_backend.assistant.service import AssistantFactory
 from fai_backend.auth.router import router as auth_router
 from fai_backend.config import settings
 from fai_backend.dependencies import get_project_user
 from fai_backend.documents.routes import router as documents_router
 from fai_backend.framework.frontend import get_frontend_environment
-from fai_backend.llm.impl.rag_wrapper import RAGWrapper
-from fai_backend.llm.models import LLMMessage, LLMDataPacket
-from fai_backend.llm.protocol import ILLMStreamProtocol
-from fai_backend.llm.serializer.impl.base64 import Base64Serializer
-from fai_backend.llm.service import LLMFactory
+from fai_backend.assistant.models import LLMClientChatMessage
+from fai_backend.assistant.protocol import ILLMProtocol
+from fai_backend.serializer.impl.base64 import Base64Serializer
 from fai_backend.logger.console import console
 from fai_backend.middleware import remove_trailing_slash
 from fai_backend.phrase import phrase as _, set_language
@@ -27,7 +26,6 @@ from fai_backend.qaf.routes import router as qaf_router
 from fai_backend.schema import ProjectUser
 from fai_backend.setup import setup_db, setup_project
 from fai_backend.vector.routes import router as vector_router
-from fai_backend.assistant.routes import router as templates_router
 
 
 @asynccontextmanager
@@ -62,7 +60,7 @@ frontend = get_frontend_environment(settings.ENV_MODE)
 frontend.configure(app)
 
 
-async def event_source_llm_generator(question: str, llm: ILLMStreamProtocol):
+async def event_source_llm_generator(question: str, llm: ILLMProtocol):
     serializer = Base64Serializer()
     stream = await llm.create()
 
@@ -73,7 +71,7 @@ async def event_source_llm_generator(question: str, llm: ILLMStreamProtocol):
             if output.final:
                 yield ServerSentEvent(
                     event="message",
-                    data=serializer.serialize(LLMMessage(
+                    data=serializer.serialize(LLMClientChatMessage(
                         date=datetime.now(),
                         source="Chat AI",
                         content=output.data
@@ -81,21 +79,12 @@ async def event_source_llm_generator(question: str, llm: ILLMStreamProtocol):
                 )
         yield ServerSentEvent(
             event="message_end",
-            data=serializer.serialize(LLMMessage(
+            data=serializer.serialize(LLMClientChatMessage(
                 date=datetime.now(),
             ))
         )
 
     return EventSourceResponse(generator())
-
-
-@app.get('/api/chat-stream')
-async def chat_stream(question: str, document: str | None = None):
-    print(f"/chat-stream {document=} {question=}")
-    llm = LLMFactory.get()
-    if document:
-        llm = RAGWrapper(question, llm, document)
-    return await event_source_llm_generator(question, llm)
 
 
 @app.get('/api/assistant-stream/{project}/{assistant}')
