+++ fai-rag-app/fai-backend/fai_backend/serializer/protocol.py
